Update abstract processing logic

+++ core/components/com_publications/admin/controllers/items.php
@@ -699,18 +699,6 @@ class Items extends AdminController
 		$metadata       = '';
 		$activity       = '';
 
-		$db = \App::get('db');
-		$db->setQuery("select params
-		               from #__extensions
-		               where name = 'Projects - Publications'");
-		$result = $db->loadRow();
-		$params = isset($result[0]) ? json_decode($result[0]) : null;
-
-		if (!!$params && isset($params->new_pubs) && !$params->new_pubs)
-		{
-		  $abstract     = htmlspecialchars(\Hubzero\Utility\Sanitize::clean($abstract));
-		}
-
 		// Save publication record
 		$this->model->publication->alias    = trim(Request::getString('alias', '', 'post'));
 		$this->model->publication->category = trim(Request::getInt('category', 0, 'post'));
@@ -760,9 +748,21 @@ class Items extends AdminController
 			}
 		}
 
+		$db = \App::get('db');
+		$db->setQuery("select params
+		               from #__extensions
+		               where folder = 'projects' and element = 'publications'");
+		$result = $db->loadRow();
+		$params = isset($result[0]) ? json_decode($result[0]) : null;
+
+		if (!!$params && isset($params->new_pubs) && !$params->new_pubs)
+		{
+			$abstract = htmlspecialchars(\Hubzero\Utility\Sanitize::clean($abstract));
+			$this->model->version->abstract = \Hubzero\Utility\Str::truncate($abstract, 250);
+		}
+
 		// Save incoming
 		$this->model->version->title        = $title;
-		$this->model->version->abstract     = \Hubzero\Utility\Str::truncate($abstract, 250);
 		$this->model->version->description  = $description;
 		$this->model->version->metadata     = $metadata;
 		$this->model->version->release_notes= $release_notes;
